safari-534.53-branch/LayoutTests/platform/gtk/accessibility/object-with-title-expected.txt

one
two
one
two
This tests the inclusion of objects with a title attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS child.role is 'AXRole: list'
PASS child.role is 'AXRole: section'
PASS child.role is 'AXRole: list'
PASS successfullyParsed is true

TEST COMPLETE
